# idle-backend/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Max Idle API",
    "version": "1.0.0",
    "description": "Generated API contract for Max Idle backend."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "token"
        ]
      },
...
      "EmailAuthRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "BetterAuthPassthrough": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        }
      }
    },
...
  },
  "paths": {
...
    "/auth/anonymous": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Create an anonymous account",
        "responses": {
          "201": {
            "description": "Anonymous credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register with email and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailAuthRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Better Auth response payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetterAuthPassthrough"
                }
              }
            }
          },
          "400": {
            "description": "Invalid registration payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login with email and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Better Auth response payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetterAuthPassthrough"
                }
              }
            }
          },
          "400": {
            "description": "Invalid login payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout current session",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Better Auth response payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetterAuthPassthrough"
                }
              }
            }
          },
          "204": {
            "description": "Logged out successfully"
          }
        }
      }
    },
...
  }
}